+++ packages/ui/src/components/common/TokenSelector/ChainFilter.tsx
@@ -1,18 +1,16 @@
 import { type FC, useState } from 'react'
-import { type Chain } from 'viem/chains'
 import { Dropdown, DropdownMenuItem } from '../../primitives/Dropdown.js'
 import { Button, Flex, Text } from '../../primitives/index.js'
 import ChainIcon from '../../primitives/ChainIcon.js'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { faChevronDown } from '@fortawesome/free-solid-svg-icons/faChevronDown'
-import type { RelayChain } from '@reservoir0x/relay-sdk'
+import type { ChainVM, RelayChain } from '@reservoir0x/relay-sdk'
 import AllChainsLogo from '../../../img/AllChainsLogo.js'
 import { TagPill } from './TagPill.js'
 
 export type ChainFilterValue =
-  | Chain
   | RelayChain
-  | { id: undefined; name: string }
+  | { id: undefined; name: string; vmType?: ChainVM }
 
 type Props = {
   options: ChainFilterValue[]