+++ packages/ui/src/utils/getTxBlockExplorerUrl.ts
@@ -1,7 +1,10 @@
 import type { RelayChain } from '@reservoir0x/relay-sdk'
 import getChainBlockExplorerUrl from './getChainBlockExplorerUrl.js'
 
-const appendQueryParams = (url: string, params: Record<string, any>) => {
+const appendQueryParams = (
+  url: string,
+  params?: Record<string, any> | null
+) => {
   if (params) {
     const queryParams = new URLSearchParams(params)
     return `${url}?${queryParams.toString()}`
@@ -18,12 +21,15 @@ export const getTxBlockExplorerUrl = (
   let blockExplorerUrl = getChainBlockExplorerUrl(chainId, chains)
 
   if (txHash) {
-    blockExplorerUrl = `${blockExplorerUrl}/tx/${txHash}`
+    if (chain?.explorerPaths?.transaction) {
+      blockExplorerUrl = `${blockExplorerUrl}${chain.explorerPaths.transaction.replace('{TX_HASH}', txHash)}`
+    } else {
+      blockExplorerUrl = `${blockExplorerUrl}/tx/${txHash}`
+    }
   }
 
   blockExplorerUrl = appendQueryParams(
     blockExplorerUrl,
-    //@ts-ignore: TODO until we update the sdk
     chain?.explorerQueryParams
   )